+++ src/types/mod.rs
@@ -196,7 +196,6 @@ macro_rules! pyobject_native_type_info(
     ($name: ty, $layout: path, $typeobject: expr,
      $module: expr $(, $checkfunction:path)? $(;$generics: ident)*) => {
         unsafe impl<$($generics,)*> $crate::type_object::PyTypeInfo for $name {
-            type Type = ();
             type BaseType = $crate::PyAny;
             type Layout = $layout;
             type BaseLayout = $crate::ffi::PyObject;